EnsureEmptyDirectory should recursively set writable perms prior to delete
